Fix deletion blacklist, indel gate, and quality array dtype bugs
- Fix blacklist range to exclude the deletion anchor so deletions
  don't remove themselves during the cleanup pass
- Fix quality array concatenation to use dtype=int so deletion errors
  (which produce an empty quality slice) don't promote the array to
  float64 and break the chr() call in finalize_read_and_write
- Remove dead num_errors capture from finalize_read_and_write return value
- Add tests for indel error paths, errors_per_read gate logic, and
  score clamping in TraditionalQualityModel

+++ neat/read_simulator/utils/read.py
@@ -71,7 +71,7 @@ class Read:
         self.read_sequence: Seq = Seq("")  # initialize to empty sequence
         self.errors: list[ErrorContainer] = []  # initialize
         self.mutations: dict[int, list] = {}  # initialize
-        self.quality_array: np.ndarray = np.zeros(self.run_read_length)  # this will have the correct memory length
+        self.quality_array: np.ndarray = np.zeros(self.run_read_length, dtype=int)  # this will have the correct memory length
         self.mapping_quality: int = 0  # initialize at 0
         self.read_quality_string: str = ""  # This will hold the read quality string
         self.num_ns = 0
@@ -161,7 +161,7 @@ class Read:
         # Replace the given quality score with the new one
         self.quality_array = \
             np.concatenate((self.quality_array[:location],
-                            np.array(new_quality_score),
+                            np.array(new_quality_score, dtype=int),
                             self.quality_array[location+ref_length:]))
 
     def apply_errors(self, quality_model: TraditionalQualityModel):
